+++ subliminal/providers/__init__.py
@@ -1,7 +1,19 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 import babelfish
+import pkg_resources
+import logging
+from ..exceptions import ProviderNotAvailable, InvalidSubtitle
 from ..video import Episode, Movie
+
+
+logger = logging.getLogger(__name__)
+
+#: Entry point for the providers
+PROVIDER_ENTRY_POINT = 'subliminal.providers'
+
+#: Available provider names
+PROVIDERS = {entry_point.name.decode('utf-8') for entry_point in pkg_resources.iter_entry_points(PROVIDER_ENTRY_POINT)}
 
 
 class Provider(object):
@@ -29,7 +41,7 @@ class Provider(object):
         self.initialize()
         return self
 
-    def __exit__(self, *args):
+    def __exit__(self, type, value, traceback):  # @ReservedAssignment
         self.terminate()
 
     def initialize(self):
@@ -114,12 +126,11 @@ class Provider(object):
         raise NotImplementedError
 
     def download_subtitle(self, subtitle):
-        """Download the `subtitle`
+        """Download the `subtitle` an fill its :attr:`~subliminal.subtitle.Subtitle.content` attribute with
+        subtitle's text
 
         :param subtitle: subtitle to download
         :type subtitle: :class:`~subliminal.subtitle.Subtitle`
-        :return: the subtitle text
-        :rtype: string
         :raise: :class:`~subliminal.exceptions.ProviderNotAvailable` if the provider is unavailable
         :raise: :class:`~subliminal.exceptions.InvalidSubtitle` if the downloaded subtitle is invalid
         :raise: :class:`~subliminal.exceptions.ProviderError` if something unexpected occured
@@ -129,3 +140,128 @@ class Provider(object):
 
     def __repr__(self):
         return '<%s [%r]>' % (self.__class__.__name__, self.video_types)
+
+
+def get_provider(name):
+    """Get a :class:`Provider` class by its name from the :data:`PROVIDER_ENTRY_POINT` entry point
+
+    :param string name: name of the provider
+    :return: the matching :class:`Provider`
+    :rtype: :class:`Provider` class
+    :raise: ValueError if the :class:`Provider` is not found
+
+    """
+    for entry_point in pkg_resources.iter_entry_points(PROVIDER_ENTRY_POINT):
+        if entry_point.name.decode('utf-8') == name:
+            return entry_point.load()
+    raise ValueError('Provider %r not found' % name)
+
+
+class ProviderManager(object):
+    """A :class:`ProviderManager` makes the :class:`Provider` API available for a set of :class:`Provider`
+
+    The :class:`ProviderManager` supports the ``with`` statement to :meth:`terminate` the providers
+
+    :param providers: providers to use, if not all
+    :type providers: list of string or None
+    :param provider_configs: configuration for providers
+    :type provider_configs: dict of provider name => provider constructor kwargs or None
+
+    """
+    def __init__(self, providers=None, provider_configs=None):
+        self.provider_configs = provider_configs or {}
+        self.providers = {provider_name: get_provider(provider_name) for provider_name in (providers or PROVIDERS)}
+        self.initialized_providers = {}
+        self.discarded_providers = set()
+
+    def __enter__(self):
+        return self
+
+    def __exit__(self, type, value, traceback):  # @ReservedAssignment
+        self.terminate()
+
+    def get_initialized_provider(self, name):
+        """Get a :class:`Provider` by name, initializing it if necessary
+
+        :param string name: name of the provider
+        :return: the initialized provider
+        :rtype: :class:`Provider`
+
+        """
+        if name in self.initialized_providers:
+            return self.initialized_providers[name]
+        provider = self.providers[name](**self.provider_configs.get(name, {}))
+        provider.initialize()
+        self.initialized_providers[name] = provider
+        return provider
+
+    def list_subtitles(self, video, languages):
+        """List subtitles for `video` with the given `languages`
+
+        :param video: video to list subtitles for
+        :type video: :class:`~subliminal.video.Video`
+        :param languages: languages of subtitles to search for
+        :type languages: set of :class:`babelfish.Language`
+        :return: found subtitles
+        :rtype: list of :class:`~subliminal.subtitle.Subtitle`
+
+        """
+        subtitles = []
+        for provider_name, provider_class in self.providers.items():
+            if not provider_class.check(video):
+                logger.info('Skipping provider %r: not a valid video', provider_name)
+                continue
+            provider_languages = provider_class.languages & languages - video.subtitle_languages
+            if not provider_languages:
+                logger.info('Skipping provider %r: no language to search for', provider_name)
+                continue
+            if provider_name in self.discarded_providers:
+                logger.debug('Skipping discarded provider %r', provider_name)
+                continue
+            try:
+                provider = self.get_initialized_provider(provider_name)
+                logger.info('Listing subtitles with provider %r and languages %r', provider_name, provider_languages)
+                provider_subtitles = provider.list_subtitles(video, provider_languages)
+                logger.info('Found %d subtitles', len(provider_subtitles))
+                subtitles.extend(provider_subtitles)
+            except ProviderNotAvailable:
+                logger.warning('Provider %r is not available, discarding it', provider_name)
+                self.discarded_providers.add(provider_name)
+            except:
+                logger.exception('Unexpected error in provider %r', provider_name)
+        return subtitles
+
+    def download_subtitle(self, subtitle):
+        """Download a subtitle
+
+        :param subtitle: subtitle to download
+        :type subtitle: :class:`~subliminal.subtitle.Subtitle`
+        :return: ``True`` if the subtitle has been successfully downloaded, ``False`` otherwise
+        :rtype: bool
+
+        """
+        if subtitle.provider_name in self.discarded_providers:
+            logger.debug('Discarded provider %r', subtitle.provider_name)
+            return False
+        try:
+            provider = self.get_initialized_provider(subtitle.provider_name)
+            provider.download_subtitle(subtitle)
+            return True
+        except ProviderNotAvailable:
+            logger.warning('Provider %r is not available, discarding it', subtitle.provider_name)
+            self.discarded_providers.add(subtitle.provider_name)
+        except InvalidSubtitle:
+            logger.warning('Invalid subtitle')
+        except:
+            logger.exception('Unexpected error in provider %r', subtitle.provider_name)
+        return False
+
+    def terminate(self):
+        """Terminate all the initialized providers"""
+        for (provider_name, provider) in self.initialized_providers.items():
+            try:
+                provider.terminate()
+            except ProviderNotAvailable:
+                logger.warning('Provider %r is not available, unable to terminate', provider_name)
+            except:
+                logger.exception('Unexpected error in provider %r', provider_name)