+++ subliminal/providers/opensubtitles.py
@@ -156,4 +156,4 @@ class OpenSubtitlesProvider(Provider):
         subtitle_text = subtitle_bytes.decode(charade.detect(subtitle_bytes)['encoding'], 'replace')
         if not is_valid_subtitle(subtitle_text):
             raise InvalidSubtitle
-        return subtitle_text
+        subtitle.content = subtitle_text