+++ subliminal/providers/tvsubtitles.py
@@ -172,4 +172,4 @@ class TVsubtitlesProvider(Provider):
         subtitle_text = subtitle_bytes.decode(charade.detect(subtitle_bytes)['encoding'], 'replace')
         if not is_valid_subtitle(subtitle_text):
             raise InvalidSubtitle
-        return subtitle_text
+        subtitle.content = subtitle_text