+++ subliminal/video.py
@@ -65,6 +65,10 @@ class Video(object):
             return Movie.fromguess(name, guess)
         raise ValueError('The guess must be an episode or a movie guess')
 
+    @classmethod
+    def fromname(cls, name):
+        cls.fromguess(os.path.split(name)[1], guessit.guess_file_info(name, 'autodetect'))
+
     def __repr__(self):
         return '<%s [%r]>' % (self.__class__.__name__, self.name)
 
@@ -109,6 +113,10 @@ class Episode(Video):
                    video_codec=guess.get('videoCodec'), audio_codec=guess.get('audioCodec'),
                    title=guess.get('title'))
 
+    @classmethod
+    def fromname(cls, name):
+        cls.fromguess(os.path.split(name)[1], guessit.guess_episode_info(name))
+
     def __repr__(self):
         return '<%s [%r, %rx%r]>' % (self.__class__.__name__, self.series, self.season, self.episode)
 
@@ -141,6 +149,10 @@ class Movie(Video):
         return cls(name, guess['title'], release_group=guess.get('releaseGroup'), resolution=guess.get('screenSize'),
                    video_codec=guess.get('videoCodec'), audio_codec=guess.get('audioCodec'),
                    year=guess.get('year'))
+
+    @classmethod
+    def fromname(cls, name):
+        cls.fromguess(os.path.split(name)[1], guessit.guess_movie_info(name))
 
     def __repr__(self):
         if self.year is None: